Refactored PathManager
Renamed PathManager to ModelIO. Refactored so a ModelIO instance is passed to models on init instead of a crufty list of parameters. Noticed that number of tiles was not being calculated correctly when jitterring tiles, which reduced training set size on individual epochs. Added misc.py module for miscellaneous helpers (so far, a platform-independent clear screen). Improved gitignore of images