bordering on insanity
Increased border size so that edge effects have less chance of propagating into the real pixels of the tile. Changed to model.predict because that's more appropriate for processing an array of tiles. Also made batch size smaller because larger models + full frame was causing the GPU to clog resulting in Tensorflow running out of memory and barfing.